+++ plot/viewer/src/canvases/ServicesCanvas.tsx
@@ -17,7 +17,6 @@ export function ServicesCanvas(props: SketchCanvasProps) {
       shouldDrill={shouldDrillService}
       showFoldButton={true}
       injectAnchor={true}
-      layoutAlgo="radial"
     />
   );
 }